requirements.txt
python-twitch-client
selenium
webdriver-manager
